feat: added duckduckgo search
