+++ libem/browse/engine/__init__.py
@@ -1,0 +1,2 @@
+from libem.browse.engine.google import search_wrapper as google_search
+from libem.browse.engine.duckduckgo import search_wrapper as duckduckgo_search
